Blender ui become a module instead of simple script Fix --debug-plugins command line First draft to migrate OpenAleaLab MainWindow to SplittableUi
